+++ pykechain/__about__.py
@@ -2,7 +2,7 @@
 name = 'pykechain'
 description = 'KE-chain Python SDK'
 
-version = '3.6.2'
+version = '3.6.4'
 
 author = 'KE-works BV'
 email = 'support+pykechain@ke-works.com'